# Fix a few spaceleaks