--- a/hledger-lib/Hledger/Data/Amount.hs
+++ b/hledger-lib/Hledger/Data/Amount.hs
@@ -435,7 +435,7 @@ tests_normaliseMixedAmountSquashPricesForDisplay = [
 -- rendering helper.
 sumSimilarAmountsUsingFirstPrice :: [Amount] -> Amount
 sumSimilarAmountsUsingFirstPrice [] = nullamt
-sumSimilarAmountsUsingFirstPrice as = (sum as){aprice=aprice $ head as}
+sumSimilarAmountsUsingFirstPrice as = (sumStrict as){aprice=aprice $ head as}
 
 -- -- | Sum same-commodity amounts. If there were different prices, set
 -- -- the price to a special marker indicating "various". Only used as a