This is the result of applying Neil Mitchell’s trick of limiting the stack size to hledger.

Let me break down the different changes:
- `sum`, `maximum` and `minimum` use `foldl` instead of `foldl'` which can leak memory.
- Since `asprecision` is rarely evaluated this is responsible for another spaceleak.
- Lastly `sequence` is sort of a false positive so I mostly made the change here to be able to use the trick. Looking at [benchmarks](http://www.joachim-breitner.de/blog/684-Constructing_a_list_in_a_monad_revisited) it seems like this difference list based version is slightly faster on small lists but is slower on bigger lists.

Here are the benchmark results:

**before**

```
Benchmark bench: RUNNING...
Benchmarking hledger in /home/moritz/code/haskell/hledger/baseline/hledger with criterion
benchmarking read bench/10000x1000x10.journal
time                 1.992 s    (1.990 s .. 1.995 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.987 s    (1.986 s .. 1.988 s)
std dev              1.392 ms   (0.0 s .. 1.543 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking print
time                 1.599 s    (1.583 s .. 1.607 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.612 s    (1.606 s .. 1.614 s)
std dev              5.457 ms   (0.0 s .. 6.127 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking register
time                 2.203 s    (2.161 s .. 2.261 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.159 s    (2.131 s .. 2.173 s)
std dev              24.04 ms   (0.0 s .. 24.94 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking balance
time                 209.6 ms   (204.6 ms .. 214.6 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 208.9 ms   (208.1 ms .. 211.0 ms)
std dev              1.625 ms   (24.09 μs .. 2.165 ms)
variance introduced by outliers: 14% (moderately inflated)

benchmarking stats
time                 195.6 ms   (189.5 ms .. 201.6 ms)
                     0.999 R²   (0.996 R² .. 1.000 R²)
mean                 197.3 ms   (194.6 ms .. 201.5 ms)
std dev              4.471 ms   (2.152 ms .. 6.746 ms)
variance introduced by outliers: 14% (moderately inflated)
```

**after**

```
Benchmark bench: RUNNING...
Benchmarking hledger in /home/moritz/code/haskell/hledger/hledger with criterion
benchmarking read bench/10000x1000x10.journal
time                 1.852 s    (1.817 s .. 1.903 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.846 s    (1.834 s .. 1.854 s)
std dev              11.56 ms   (0.0 s .. 13.28 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking print
time                 1.602 s    (1.589 s .. 1.609 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.607 s    (1.606 s .. 1.607 s)
std dev              352.9 μs   (0.0 s .. 407.5 μs)
variance introduced by outliers: 19% (moderately inflated)

benchmarking register
time                 2.143 s    (2.050 s .. 2.223 s)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 2.135 s    (2.121 s .. 2.146 s)
std dev              17.79 ms   (0.0 s .. 19.60 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking balance
time                 204.6 ms   (196.1 ms .. 209.5 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 204.5 ms   (202.2 ms .. 206.1 ms)
std dev              2.434 ms   (1.329 ms .. 3.665 ms)
variance introduced by outliers: 14% (moderately inflated)

benchmarking stats
time                 190.3 ms   (182.7 ms .. 195.5 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 190.9 ms   (188.6 ms .. 194.4 ms)
std dev              3.729 ms   (696.3 μs .. 5.146 ms)
variance introduced by outliers: 14% (moderately inflated)
```

Memory usage doesn’t seem to be significantly impacted (although it looks like it goes slightly down with this patch).