@simonmichael 2017-01-11
I'd like to benchmark this too, first rebasing it against master so I am comparing like with like, but by now it has two conflicts with master which are not so easy for me to resolve correctly. @cocreature, could you do the rebase ?