@cocreature 2017-01-12
@simonmichael rebased.