@simonmichael 2017-01-12
Thanks. Here's what I get:
```
$ quickbench -n4 -w hledger-master-d92d777c,hledger-413-5bc00818 'hledger -f examples/10000x1000x10.journal stats' 'hledger -f examples/10000x1000x10.journal print' 'hledger -f examples/10000x1000x10.journal balance' 'hledger -f examples/10000x1000x10.journal register'
Running 4 tests 4 times with 2 executables at 2017-01-12 12:48:04 PST:

Best times:
+--------------------------------------------++-------------------------+----------------------+
|                                            || hledger-master-d92d777c | hledger-413-5bc00818 |
+============================================++=========================+======================+
| -f examples/10000x1000x10.journal stats    ||                    2.03 |                 2.07 |
| -f examples/10000x1000x10.journal print    ||                    3.05 |                 3.12 |
| -f examples/10000x1000x10.journal balance  ||                    2.03 |                 2.21 |
| -f examples/10000x1000x10.journal register ||                    4.02 |                 3.94 |
+--------------------------------------------++-------------------------+----------------------+
```

Or using the benchmark suite:

master:
```
~/src/hledger$ stack bench hledger
hledger-1.1.98: benchmarks
Running 1 benchmarks...
Benchmark bench: RUNNING...
Benchmarking hledger in /Users/simon/src/hledger/hledger with timeit
read bench/10000x1000x10.journal        [1.63s]
print                                   [1.23s]
register                                [1.67s]
balance                                 [0.15s]
Total: 4.68s
Benchmark bench: FINISH
```

413:
```
~/src/hledger$ stack bench hledger
hledger-1.1.98: benchmarks
Running 1 benchmarks...
Benchmark bench: RUNNING...
Benchmarking hledger in /Users/simon/src/hledger/hledger with timeit
read bench/10000x1000x10.journal        [1.73s]
print                                   [1.14s]
register                                [1.72s]
balance                                 [0.15s]
Total: 4.74s
Benchmark bench: FINISH
Completed 3 action(s).
```

Or with criterion:

master:
```
~/src/hledger$ stack bench hledger --benchmark-arguments  --criterion
hledger-1.1.98: benchmarks
Running 1 benchmarks...
Benchmark bench: RUNNING...
Benchmarking hledger in /Users/simon/src/hledger/hledger with criterion
benchmarking read bench/10000x1000x10.journal
time                 1.589 s    (1.472 s .. 1.759 s)
                     0.999 R²   (0.996 R² .. 1.000 R²)
mean                 1.620 s    (1.594 s .. 1.644 s)
std dev              39.29 ms   (0.0 s .. 41.16 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking print
time                 1.091 s    (857.4 ms .. 1.385 s)
                     0.992 R²   (0.973 R² .. 1.000 R²)
mean                 1.155 s    (1.111 s .. 1.198 s)
std dev              74.08 ms   (0.0 s .. 74.49 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking register
time                 1.607 s    (1.437 s .. 1.867 s)
                     0.997 R²   (0.992 R² .. 1.000 R²)
mean                 1.679 s    (1.631 s .. 1.721 s)
std dev              67.61 ms   (0.0 s .. 72.42 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking balance
time                 172.2 ms   (162.8 ms .. 187.2 ms)
                     0.993 R²   (0.976 R² .. 1.000 R²)
mean                 172.5 ms   (169.3 ms .. 177.7 ms)
std dev              5.662 ms   (2.580 ms .. 8.277 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking stats
time                 141.8 ms   (133.0 ms .. 155.5 ms)
                     0.995 R²   (0.987 R² .. 1.000 R²)
mean                 150.7 ms   (146.1 ms .. 155.2 ms)
std dev              6.634 ms   (4.564 ms .. 8.717 ms)
variance introduced by outliers: 12% (moderately inflated)

Benchmark bench: FINISH
```

413:
```
~/src/hledger$ stack bench hledger --benchmark-arguments  --criterion
hledger-1.1.98: benchmarks
Running 1 benchmarks...
Benchmark bench: RUNNING...
Benchmarking hledger in /Users/simon/src/hledger/hledger with criterion
benchmarking read bench/10000x1000x10.journal
time                 1.692 s    (1.091 s .. 2.433 s)
                     0.978 R²   (0.932 R² .. 1.000 R²)
mean                 1.799 s    (1.710 s .. 1.879 s)
std dev              132.3 ms   (271.9 as .. 139.2 ms)
variance introduced by outliers: 20% (moderately inflated)

benchmarking print
time                 1.189 s    (1.061 s .. 1.411 s)
                     0.996 R²   (0.990 R² .. 1.000 R²)
mean                 1.116 s    (1.085 s .. 1.166 s)
std dev              44.08 ms   (0.0 s .. 47.06 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking register
time                 1.512 s    (1.097 s .. 1.994 s)
                     0.988 R²   (0.960 R² .. 1.000 R²)
mean                 1.736 s    (1.633 s .. 1.834 s)
std dev              164.6 ms   (0.0 s .. 169.2 ms)
variance introduced by outliers: 22% (moderately inflated)

benchmarking balance
time                 159.6 ms   (153.6 ms .. 164.1 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 159.9 ms   (157.5 ms .. 162.9 ms)
std dev              3.569 ms   (2.154 ms .. 5.673 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking stats
time                 150.9 ms   (144.5 ms .. 157.4 ms)
                     0.998 R²   (0.996 R² .. 1.000 R²)
mean                 160.5 ms   (156.0 ms .. 166.3 ms)
std dev              7.110 ms   (4.594 ms .. 9.727 ms)
variance introduced by outliers: 12% (moderately inflated)

Benchmark bench: FINISH
```

If anything, it seems slower, as far as I can measure ?

Some of that new code sounds nice to have, but we should probably see a real benefit before giving up the readability and familiarity and low maintenance of standard library functions. I think that also goes for INLINABLE, which we discussed on another PR.