@cocreature 2017-01-12
> Some of that new code sounds nice to have, but we should probably see a real benefit before giving up the readability and familiarity and low maintenance of standard library functions. I think that also goes for INLINABLE, which we discussed on another PR.

I disagree with that part. Adding `INLINABLE` to improve the performance is a nice to have for me. I see spaceleaks as bugs so they always should be fixed.

The differences in the benchmarks look too small to be significant. I would expect that if you run it a few times you’ll see differences in that range in both directions.